requirements.txt
Flask==3.0.0
python-docx==1.1.0
gunicorn==21.2.0
Flask-login==0.6.3
Flask-dance==7.1.0
docx2pdf==0.3.0